requirements.txt
atomicwrites==1.1.5
attrs==18.1.0
backcall==0.1.0
clear==1.0.1
coverage==4.5.1
decorator==4.3.0
Django==1.11.8
django-rest-framework==0.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
Faker==0.8.8
goodlogging==1.0.1
gunicorn==19.9.0
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
mixer==6.0.1
more-itertools==4.3.0
parso==0.3.1
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pkg-resources==0.0.0
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
py==1.5.4
Pygments==2.2.0
PyJWT==1.6.4
pytest==3.7.2
pytest-cov==2.5.1
pytest-django==3.4.1
pytest-ipdb==0.1
python-dateutil==2.7.3
pytz==2018.5
rarfile==2.7
requests==2.6.0
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
